add stack overflow protection
linking now must be done using the `flip-lld` linker wrapper (this is the
default for all the examples). This linker wrapper flips the standard memory
layout of programs so that the stack cannot overwrite unrelated memory when it
grows in an unbounded fashion. More details can be found in the blog post:
https://blog.japaric.io/stack-overflow-protection/

Stack overflows will now trigger a "data abort" (hardware) exception when memory
is exhausted. The handler for this exception will print a message to the
console (using the `Serial` API) and then reboot the SoC. An example has been
added to demo this behavior.

closes #11