remove DWARF (`.debug_*`) sections from the ELF
we are not using the debugger so the host does not make use these sections. This
reduces the size of ELF files by 33-88%, which makes transferring them over the
slow serial interface 3 to 8 times faster. As loading the program into the
target using u-boot requires copying the whole ELF into DDR RAM first, this
significantly improves interactive program loading times.

This change doesn't affect the amount of memory used by the program when it's
running.